+++ .gitignore
@@ -67,7 +67,8 @@ condaenv.*.requirements.txt
 /flagged/*
 /gfpgan/*
 /models/*
-/assets/*
+/webui/flet/uploads
+
 z_version_env.tmp
 scripts/bridgeData.py
 /user_data/*
