+++ configs/webui/webui_flet.yaml
@@ -33,6 +33,7 @@
 #       -option_list
 #   slider
 #       -value
+#       -value_type  'int', 'float'
 #       -min
 #       -max
 #       -step
@@ -56,12 +57,14 @@ webui_page:
     default_text_size:
         display: slider
         value: !!int '20'
+        value_type: 'int'
         min: !!int '10'
         max: !!int '32'
         step: !!float '2.0'
     max_message_history:
         display: slider
         value: !!int '20'
+        value_type: 'int'
         min: !!int '1'
         max: !!int '100'
         step: !!int '1'
